Extract board impl orchestration and MP emissions
