+++ crates/fstart-codegen/src/stage_gen/board_gen.rs
@@ -42,13 +42,10 @@
 //!    trait method.
 
 use proc_macro2::TokenStream;
-use quote::{format_ident, quote};
-
 use fstart_device_registry::{DriverInstance, Service};
-use fstart_types::{BoardConfig, BootMedium, Capability, DeviceConfig, DeviceNode, Platform};
+use fstart_types::{BoardConfig, BootMedium, Capability, DeviceConfig, DeviceNode};
 
-use super::tokens::hex_addr;
-
+mod board_impl;
 mod boot_media;
 mod caps_tables;
 mod fdt;
@@ -56,22 +53,16 @@ mod init_caps;
 mod lifecycle;
 mod logger;
 mod model;
+mod mp;
 mod payload;
 mod phases;
+mod security;
 mod state;
 mod sunxi;
 
-use boot_media::{anchor_bytes_stmt, match_boot_media};
-use caps_tables::{acpi_load_body, acpi_prepare_body, memory_detect_body, smbios_prepare_body};
-use fdt::{fdt_prepare_body, return_to_fel_body, stage_load_body};
-use init_caps::{dram_init_body, late_driver_init_body, pci_init_body};
-use lifecycle::{init_all_devices_body, init_device_body};
-use logger::install_logger_body;
-use model::{device_provides, BoardCtx};
-use payload::payload_load_body;
-use phases::{phase_init_body, PhaseSpec};
+use board_impl::emit_board_impl;
+use model::BoardCtx;
 use state::{emit_adapter_new, emit_adapter_struct};
-use sunxi::{boot_media_select_body, load_next_stage_body};
 
 // =======================================================================
 // Public entry point
@@ -207,484 +198,6 @@ fn compute_excluded_indices(
 // `impl Board for _BoardDevices`
 // =======================================================================
 
-/// Emit the `impl fstart_stage_runtime::Board for _BoardDevices` block.
-///
-/// Most methods are small trampolines from the generic stage executor into
-/// the concrete drivers and capability helpers selected by the board RON.
-fn emit_board_impl(platform: Platform, ctx: &BoardCtx<'_>) -> TokenStream {
-    let jump_with_handoff_body = match platform {
-        Platform::X86_64 => quote! {
-            // Unsupported on x86_64; should never be reached — no x86
-            // board uses a LoadNextStage capability today.  If it ever
-            // does, a platform-crate helper needs to land first.
-            let _ = (entry, handoff_addr);
-            fstart_platform::halt()
-        },
-        _ => quote! { fstart_platform::jump_to_with_handoff(entry, handoff_addr) },
-    };
-
-    let sig_verify_body = sig_verify_body(ctx);
-    let fdt_prepare_body = fdt_prepare_body(platform, ctx);
-    let install_logger_body = install_logger_body(ctx);
-    let stage_load_body = stage_load_body(ctx);
-    let return_to_fel_body = return_to_fel_body(platform, ctx);
-    let pci_init_body = pci_init_body(ctx);
-    let dram_init_body = dram_init_body(ctx);
-    let pre_console_init_body = phase_init_body(
-        ctx,
-        PhaseSpec::new(
-            Service::PreConsoleInit,
-            "PreConsoleInit",
-            "pre_console_init",
-        ),
-    );
-    let early_init_body = phase_init_body(
-        ctx,
-        PhaseSpec::new(Service::EarlyInit, "EarlyInit", "early_init"),
-    );
-    let stage_local_init_body = phase_init_body(
-        ctx,
-        PhaseSpec::new(
-            Service::StageLocalInit,
-            "StageLocalInit",
-            "stage_local_init",
-        ),
-    );
-    let post_dram_init_body = phase_init_body(
-        ctx,
-        PhaseSpec::new(Service::PostDramInit, "PostDramInit", "post_dram_init"),
-    );
-    let finalize_init_body = phase_init_body(
-        ctx,
-        PhaseSpec::new(Service::FinalizeInit, "FinalizeInit", "finalize_init"),
-    );
-
-    let acpi_load_body = acpi_load_body(ctx);
-    let memory_detect_body = memory_detect_body(ctx);
-    let acpi_prepare_body = acpi_prepare_body(ctx);
-    let smbios_prepare_body = smbios_prepare_body(ctx);
-    let mp_init_body = mp_init_body(ctx);
-    let boot_media_select_body = boot_media_select_body(ctx);
-    let load_next_stage_body = load_next_stage_body(ctx);
-    let payload_load_body = payload_load_body(platform, ctx);
-    let init_device_body = init_device_body(ctx);
-    let init_all_devices_body = init_all_devices_body(ctx);
-    let late_driver_init_body = late_driver_init_body(ctx);
-    let boot_media_context_publish = if ctx.stage.uses_ffs {
-        let anchor = anchor_bytes_stmt();
-        quote! {
-            if device.is_none() {
-                #anchor
-                fstart_services::ffs_context::set_memory_mapped(_anchor_bytes, offset, size);
-            }
-        }
-    } else {
-        quote! {}
-    };
-
-    quote! {
-        #[allow(dead_code, unused_variables)]
-        impl fstart_stage_runtime::Board for _BoardDevices {
-            // ----- Device lifecycle -------------------------------------
-
-            fn init_device(
-                &mut self,
-                id: fstart_types::DeviceId,
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #init_device_body
-            }
-
-            fn init_all_devices(
-                &mut self,
-                skip: &fstart_stage_runtime::DeviceMask,
-                gated: &fstart_stage_runtime::DeviceMask,
-            ) {
-                #init_all_devices_body
-            }
-
-            // ----- Logging ---------------------------------------------
-
-            unsafe fn install_logger(&self, id: fstart_types::DeviceId) {
-                #install_logger_body
-            }
-
-            // ----- Capability trampolines ------------------------------
-
-            fn memory_init(&self) {
-                fstart_capabilities::memory_init();
-            }
-
-            fn late_driver_init_complete(&mut self, count: usize) {
-                #late_driver_init_body
-                fstart_capabilities::late_driver_init_complete(count);
-            }
-
-            fn sig_verify(&self) {
-                #sig_verify_body
-            }
-
-            fn fdt_prepare(&self) {
-                #fdt_prepare_body
-            }
-
-            fn payload_load(&self) -> ! {
-                #payload_load_body
-            }
-
-            fn stage_load(&self, next_stage: &str) -> ! {
-                #stage_load_body
-            }
-
-            fn acpi_prepare(&mut self) {
-                #acpi_prepare_body
-            }
-
-            fn smbios_prepare(&self) {
-                #smbios_prepare_body
-            }
-
-            fn mp_init(
-                &mut self,
-                cpu_model: &str,
-                num_cpus: u16,
-                smm: bool,
-            ) -> Result<(), fstart_stage_runtime::RuntimeError> {
-                #mp_init_body
-            }
-
-            fn pre_console_init(
-                &mut self,
-                ids: &[fstart_types::DeviceId],
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #pre_console_init_body
-            }
-
-            fn early_init(
-                &mut self,
-                ids: &[fstart_types::DeviceId],
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #early_init_body
-            }
-
-            fn stage_local_init(
-                &mut self,
-                ids: &[fstart_types::DeviceId],
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #stage_local_init_body
-            }
-
-            fn post_dram_init(
-                &mut self,
-                ids: &[fstart_types::DeviceId],
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #post_dram_init_body
-            }
-
-            fn finalize_init(
-                &mut self,
-                ids: &[fstart_types::DeviceId],
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #finalize_init_body
-            }
-
-            fn dram_init(
-                &mut self,
-                id: fstart_types::DeviceId,
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #dram_init_body
-            }
-
-            fn pci_init(
-                &mut self,
-                id: fstart_types::DeviceId,
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #pci_init_body
-            }
-
-            fn acpi_load(
-                &mut self,
-                id: fstart_types::DeviceId,
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #acpi_load_body
-            }
-
-            fn memory_detect(
-                &mut self,
-                id: fstart_types::DeviceId,
-            ) -> Result<(), fstart_services::device::DeviceError> {
-                #memory_detect_body
-            }
-
-            fn return_to_fel(&self) -> ! {
-                #return_to_fel_body
-            }
-
-            // ----- Boot media ------------------------------------------
-
-            fn boot_media_select(
-                &mut self,
-                candidates: &[fstart_stage_runtime::BootMediaCandidate],
-            ) -> Option<fstart_types::DeviceId> {
-                #boot_media_select_body
-            }
-
-            fn boot_media_static(
-                &mut self,
-                device: Option<fstart_types::DeviceId>,
-                offset: u64,
-                size: u64,
-            ) {
-                // Records the executor-provided selection on `self`
-                // for later FFS-using trampolines (`sig_verify`,
-                // `payload_load`, `stage_load`) to reconstruct the
-                // concrete `impl BootMedia`.
-                //
-                // Note: the executor already called `init_device(id)`
-                // for `Some(id)` before calling us, so by the time
-                // `sig_verify` etc. dereference `self.<name>`, the
-                // device is constructed.
-                self._boot_media =
-                    fstart_stage_runtime::BootMediaState::from_static(device, offset, size);
-                #boot_media_context_publish
-            }
-
-            fn load_next_stage(&mut self, next_stage: &str) -> ! {
-                #load_next_stage_body
-            }
-
-            // ----- Platform primitives ---------------------------------
-
-            fn halt(&self) -> ! { fstart_platform::halt() }
-
-            fn jump_to(&self, entry: u64) -> ! {
-                fstart_platform::jump_to(entry)
-            }
-
-            fn jump_to_with_handoff(&self, entry: u64, handoff_addr: usize) -> ! {
-                #jump_with_handoff_body
-            }
-        }
-    }
-}
-
-// =======================================================================
-// Capability body helpers
-// =======================================================================
-
-fn mp_init_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    let uses_mp = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|cap| matches!(cap, Capability::MpInit { .. }));
-    if !uses_mp {
-        return quote! {
-            let _ = (cpu_model, num_cpus, smm);
-            Ok(())
-        };
-    }
-
-    let uses_smm = ctx
-        .stage
-        .capabilities
-        .iter()
-        .any(|cap| matches!(cap, Capability::MpInit { smm: true, .. }));
-    let smm_image_expr = if uses_smm {
-        quote! { if smm { Some(FSTART_SMM_IMAGE) } else { None } }
-    } else {
-        quote! { None }
-    };
-
-    let explicit_smm_provider = ctx.stage.capabilities.iter().find_map(|cap| match cap {
-        Capability::MpInit {
-            smm: true,
-            smm_provider: Some(provider),
-            ..
-        } => Some(provider.as_str()),
-        _ => None,
-    });
-    let smm_provider = if let Some(provider) = explicit_smm_provider {
-        enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-            .find(|idx| ctx.devices[*idx].name.as_str() == provider)
-    } else {
-        enabled_indices(ctx.devices, ctx.instances, ctx.excluded)
-            .find(|idx| device_provides(ctx, *idx, Service::SmmOps))
-    };
-
-    let mp_microcode_enabled = matches!(
-        ctx.config.microcode.as_ref(),
-        Some(fstart_types::board::MicrocodeConfig::Intel(config)) if config.mp
-    );
-    let microcode_expr = mp_microcode_enabled
-        .then(|| ctx.stage.capabilities.iter().find_map(|cap| match cap {
-            Capability::BootMedia(BootMedium::MemoryMapped { base, .. }) => Some(*base),
-            _ => None,
-        }))
-        .flatten()
-        .map(|base| {
-            let base_lit = hex_addr(base);
-            let anchor_stmt = anchor_bytes_stmt();
-            quote! {
-                {
-                    #anchor_stmt
-                    let anchor = unsafe { fstart_ffs::FfsReader::read_anchor_volatile(_anchor_bytes) }
-                        .ok();
-                    anchor.and_then(|anchor| {
-                        if anchor.microcode_offset == 0 || anchor.microcode_size == 0 {
-                            None
-                        } else {
-                            let addr = (#base_lit as usize).saturating_add(anchor.microcode_offset as usize);
-                            // SAFETY: xtask patched the anchor with a range inside the
-                            // memory-mapped FFS image declared by this stage's BootMedia.
-                            Some(unsafe {
-                                core::slice::from_raw_parts(addr as *const u8, anchor.microcode_size as usize)
-                            })
-                        }
-                    })
-                }
-            }
-        })
-        .unwrap_or_else(|| quote! { None });
-
-    let smm_ops_expr = if let Some(idx) = smm_provider {
-        let field = format_ident!("{}", ctx.devices[idx].name.as_str());
-        quote! {
-            if smm {
-                Some(
-                    self.#field
-                        .as_ref()
-                        .ok_or(fstart_stage_runtime::RuntimeError::Failed)?
-                        as &dyn fstart_mp::SmmOps,
-                )
-            } else {
-                None
-            }
-        }
-    } else {
-        quote! {
-            if smm {
-                fstart_log::error!("mp: SMM requested but this board has no SmmOps provider");
-                return Err(fstart_stage_runtime::RuntimeError::Failed);
-            } else {
-                None
-            }
-        }
-    };
-
-    quote! {
-        let smm_ops: Option<&dyn fstart_mp::SmmOps> = #smm_ops_expr;
-        let smm_image: Option<&[u8]> = #smm_image_expr;
-        let microcode_blob: Option<&'static [u8]> = #microcode_expr;
-
-        if cpu_model == "generic-x86" || cpu_model == "qemu-x86" || cpu_model == "qemu" {
-            let cpu_ops = fstart_mp::GenericX86CpuOps;
-            let config = fstart_mp::MpConfig {
-                cpu_ops: &cpu_ops,
-                smm: smm_ops,
-                smm_image,
-                num_cpus,
-            };
-            fstart_mp::mp_init(&config)
-                .map(|_| ())
-                .map_err(|_| fstart_stage_runtime::RuntimeError::Failed)
-        } else if cpu_model == "core2" || cpu_model == "6fx" {
-            #[cfg(feature = "core2-cpu")]
-            {
-                let cpu_ops = fstart_cpu_intel::core2_cpu::Core2CpuOps::new(0x0500, microcode_blob);
-                let config = fstart_mp::MpConfig {
-                    cpu_ops: &cpu_ops,
-                    smm: smm_ops,
-                    smm_image,
-                    num_cpus,
-                };
-                fstart_mp::mp_init(&config)
-                    .map(|_| ())
-                    .map_err(|_| fstart_stage_runtime::RuntimeError::Failed)
-            }
-            #[cfg(not(feature = "core2-cpu"))]
-            {
-                fstart_log::error!("mp: Core 2 CPU ops feature is not enabled");
-                Err(fstart_stage_runtime::RuntimeError::Failed)
-            }
-        } else if cpu_model == "pineview" || cpu_model == "106cx" {
-            #[cfg(feature = "pineview-cpu")]
-            {
-                let cpu_ops = fstart_cpu_intel::pineview::PineviewCpuOps::with_microcode(0x0500, microcode_blob);
-                let config = fstart_mp::MpConfig {
-                    cpu_ops: &cpu_ops,
-                    smm: smm_ops,
-                    smm_image,
-                    num_cpus,
-                };
-                fstart_mp::mp_init(&config)
-                    .map(|_| ())
-                    .map_err(|_| fstart_stage_runtime::RuntimeError::Failed)
-            }
-            #[cfg(not(feature = "pineview-cpu"))]
-            {
-                fstart_log::error!("mp: Pineview CPU ops feature is not enabled");
-                Err(fstart_stage_runtime::RuntimeError::Failed)
-            }
-        } else {
-            fstart_log::error!("mp: unsupported CPU model '{}'; no CpuOps provider", cpu_model);
-            Err(fstart_stage_runtime::RuntimeError::Failed)
-        }
-    }
-}
-
-/// Emit the body of `Board::sig_verify`.
-///
-/// For stages that use FFS (`ctx.stage.uses_ffs == true`), generates:
-///
-/// 1. A volatile-read of `FSTART_ANCHOR` as a `&[u8]`.
-/// 2. A match on `self._boot_media` that reconstructs the concrete
-///    boot-media type and delegates to `fstart_capabilities::sig_verify`.
-/// 3. One per-device arm for each enabled `BlockDevice` provider in
-///    the board; no-match halts.
-///
-/// For non-FFS stages, emits a `todo!()` — the trampoline is dead
-/// code (no `SigVerify` capability, so the executor never calls it)
-/// and referencing `FSTART_ANCHOR` or the `MemoryMapped` /
-/// `BlockDeviceMedia` types would break compilation since they are
-/// not imported in non-FFS stages.
-fn sig_verify_body(ctx: &BoardCtx<'_>) -> TokenStream {
-    if !ctx.stage.uses_ffs {
-        return quote! {
-            // No FFS-using capability in this stage, so no executor
-            // arm reaches `sig_verify`.  Keep the trait method but
-            // make it a compile-time-only placeholder.
-            todo!("board_gen::sig_verify: no FFS-using capability in this stage")
-        };
-    }
-
-    let bm_usage = quote! {
-        fstart_capabilities::sig_verify(_anchor_bytes, &_bm);
-    };
-    let none_body = quote! {
-        // No `BootMedia*` capability ran before `SigVerify`.
-        // Validation upstream forbids this; reaching here means a
-        // buggy plan.  Log and skip — consistent with old
-        // `fstart_capabilities::sig_verify` stub behaviour when the
-        // manifest is empty.
-        fstart_log::info!("sig verify: no boot media configured, skipping");
-    };
-    let anchor = anchor_bytes_stmt();
-    let match_body = match_boot_media(ctx, &bm_usage, "sig_verify", &none_body);
-
-    quote! {
-        #anchor
-        #match_body
-    }
-}
-
-/// The `let _anchor_bytes: &[u8] = ...;` preamble used by every FFS-
-/// touching trampoline body.
-///
-/// Emitted as a constant rather than a function because the token
-/// stream contains zero per-board variation.  Centralised so the
-/// safety comment lives in exactly one place.
-/// Emit the body of `Board::init_device`.
-///
 // =======================================================================
 // Helpers
 // =======================================================================
